# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Policlinic Management API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/outcome/records:
    get:
      summary: Outcome records
      responses:
        "200":
          description: Outcome records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutcomeRecord"
    post:
      summary: Create outcome record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutcomeRecordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
  /api/outcome/salaries:
    get:
      summary: Salary payments
      responses:
        "200":
          description: Salary payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalaryPayment"
    post:
      summary: Create salary payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalaryPaymentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
...
components:
  schemas:
...
    OutcomeRecord:
      type: object
      properties:
        id:
          type: integer
        expense_date:
          type: string
          format: date
        amount:
          type: number
        description:
          type: string
        vendor:
          type: string
        category:
          type: string
    CreateOutcomeRecordRequest:
      type: object
      properties:
        category_id:
          type: integer
        amount:
          type: number
        expense_date:
          type: string
          format: date
        description:
          type: string
        vendor:
          type: string
    SalaryPayment:
      type: object
      properties:
        id:
          type: integer
        payment_date:
          type: string
          format: date
        amount:
          type: number
        note:
          type: string
        staff:
          $ref: "#/components/schemas/StaffMember"
    CreateSalaryPaymentRequest:
      type: object
      properties:
        staff_id:
          type: integer
        amount:
          type: number
        payment_date:
          type: string
          format: date
        note:
          type: string
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        email:
          type: string
        bio:
          type: string
        base_salary:
          type: number
        last_paid_at:
          type: string
          format: date
          nullable: true
        total_revenue:
          type: number
        is_active:
          type: boolean
        role:
          type: string
...